refactor: extract games_lifecycle.py from games.py
Move save/restore/resume/update/end handlers + _VALID_TRANSITIONS
to games_lifecycle.py (371 lines). Update monkeypatch path in
test_resume_stale_summary to target new module.

games.py: 1212 → 865 lines (-347, -28.6%)
Cumulative: 1610 → 865 (games.py) + 79 + 374 + 371 (extracted)
